Colocate tools templates
