+++ controllers/tools/views/feedback.njk
@@ -1,4 +1,4 @@
-{% extends "layouts/tools.njk" %}
+{% extends "./layout.njk" %}
 {% block title %}Feedback results{% endblock %}
 
 {% block content %}
