+++ controllers/tools/views/index.njk
@@ -1,4 +1,4 @@
-{% extends "layouts/tools.njk" %}
+{% extends "./layout.njk" %}
 
 {% block title %}Tools{% endblock %}
