+++ controllers/tools/views/orders.njk
@@ -1,4 +1,4 @@
-{% extends "layouts/tools.njk" %}
+{% extends "./layout.njk" %}
 
 {% block title %}Material orders{% endblock %}
