+++ controllers/tools/views/pagelist.njk
@@ -1,4 +1,4 @@
-{% extends "layouts/tools.njk" %}
+{% extends "./layout.njk" %}
 
 {% block title %}Page Index{% endblock %}
