+++ views/user/register.njk
@@ -1,4 +1,4 @@
-{% extends "../layouts/tools.njk" %}
+{% extends "controllers/tools/views/layout.njk" %}
 {% from "./user-helpers.njk" import listErrors with context %}
 
 {% block title %}Register{% endblock %}
